+++ solr/cli/CHANGES.md
@@ -10,3 +10,8 @@
 - `bin/solr cluster` is now `bin/solr zk cluster`
 - `bin/solr start -e *` is now `bin/solr start example *`
 - `--jettyconfig` is now `--jetty-config`
+- `solr.ulimit.checks` is now `solr.ulimit.checks.enabled`
+- `SOLR_TIP` is now `SOLR_INSTALL_DIR`
+- `SOLR_TIP` is now `SOLR_INSTALL_DIR`
+- `-DdisableAdminUI` is now `solr.admin.enabled`
+- `SOLR_HEAP` is now `solr.memory.allocation.max` and `solr.memory.allocation.initial`